replaces cockroachdb treeprinter with github.com/xlab/treeprint
despite CRDB's treeprinter converting to Apache 2.0 after the Change Date,
the community reported friction on adopting `zed` because automated tools
would flag BSL incompatible with Apache 2.0

This commit replaces cockroachdb treeprinter with
github.com/xlab/treeprint. It's a mostly 1:1 migration,
except due to the design choice to always create an
implicit empty root node. I addressed that by removing that
implicit root node after the tree is populated, as
unfortunately the printer is unable to realize
we are starting the print process from
an intermediate node.