capture SIGINT and fail-fast
certain operations may take long to execute, e.g.
ReadRelationships with a very wide relation.
The current behaviour is to completely ignore
process termination via something like Control+C,
and zed will continue to block until SIGKILL or
the operation completes. It's a jarring experience.

This commit introduces a goroutine the listens
to SIGINT and cancels the context, which is also
now propagated via cobra. This makes the command-line
seem responsive when the user terminates it.