Allow specifying caveats as part of subject string
This commit changes the parsing logic for the relationship update
commands to use the tuple.ParseRel() function from SpiceDB. Doing so
allows the user to specify caveats as part of the subject argument, as
well as by the --caveat flag. If both the --caveat flag and a caveat in
the subject are specified, an error is returned.

Using the tuple.ParseRel() function will ensure that output from `zed
relationship read` can always be passed to `zed relationship
{create,touch,delete}`.

Signed-off-by: Erik Paulson <epaulson10@gmail.com>