Allow streaming relationship updates via stdin
To make Zed more performant in scripting, this commit refactors
the commands with call WriteRelationships to allow sending relationships
over stdin. When sent over stdin, relationships are newline delimted.
One should be able to pipe the output of `zed relationship read` into
`zed relationship touch`.

This functionality is similar to the `zed import` command, however this
exposes writing multiple relationships in a form more suitable for
scripting. For example, one can pipe the results of a relationship read
through sed, and then pipe that result back into `zed relationship
touch`. The alternative without this patch is to use xargs to call zed
with parameters, but this means each relationship to write will need an
entire process created and gRPC connection initiated, which is time
consuming.

Signed-off-by: Erik Paulson <epaulson10@gmail.com>