Add missing token attributes
The context token supports NoVerifyCA and CACert. When it restored the values from a keychain it wasn't mapping the values from the token read to the token returned. This prevented zed client users from benefiting from stored --no-verify-ca and --certificate-path values.

With a simple addition the values are passed from context and they no longer need to be manually set on the command line.

Signed-off-by: David Kanenwisher <dkanenwisher@bushelpowered.com>