restore: report error if one of the batches fail
if one of the batches sent fails, zed will hide
the underlying error and just report EOF.

The gRPC backend will report EOF as part of an error,
which will have to eb retrieved via CloseAndRecv().

A common error that surfaced as EOF was duplicate
relationships after reimporting the same backup file.