refactor backup import to handle serialization errors and conflicts
This commit refactors the logic on the backup command into
a restorer type. The fundamental differences with the previous logic are:
- batch conflicts are detected, and provides to handling strategies: fail, skip or touch
  when using touch, they will also be retried if needed
- serialization errors are detected, and optionally can be retried using
  WithRelationships with TOUCH semantics
- retry strategy uses a backoff

any error on bulk import will cause the stream to close, so attempts to recover
from that by retrying means retrying with another bulk import stream or
using WriteRelationships. The latter was used because it normalizes retries
as BulkImport does not support TOUCH semantics.