updates bulk-delete to use DeleteRelationships new optional-limit
former implementation of bulk-delete was designed to support
deleting large amounts of relationships in a batched way.

This commit replaces the implementation that relied on ReadRelationships
with the new native DeleteRelationships optional-limit.

The contract is kept mostly intact, except the interactive prompt is removed
as it was considered bad UX. Instead, the same limits of 1000 before the
user is prompted is retained, and instead zed will fail, indicating
the user to bump the limit with the new --optional-limit flag, or
to use --force.

The --optional-limit flag sets the corresponding `DeleteRelationship` request
field, and `--force` now sets the `OptionalAllowPartialDeletions` field. If
set to true, all relationships will be deleted in batches defined by `--optional-limit`,
which continues to default to 1000 elements.

It also adds a progress bar to provide visual feedback to the user while
deletion takes place.