introduces support for optional resource type in ResourceFilter
This introduces support for optional resource type, recently introduced
in SpiceDB 1.30.0.

since the CLI contract has ambiguity, as a tradeoff, the contract
will break by requiring clients to specify colon after
the resource type, in order to disambiguate with _only
specifying the relation_. The subject-type continues to
not require colon as suffix.

Tests were also added for missing permutations and the optional
resource id prefix.