introduce support for resuming a backup
this commit changes the backup create command
so that a canceled backup can be resumed.

A new marker file is added with the last written
bulk export cursor. This was not added to the OCF
file because OCF container is not meant for in-place
updates, but streaming append operations. The marker
is updated every time a bulk export page is successfully
written.

For convenience the command now also supports creating
a backup without a file name: in that case it will
derive the backup file name from the zed context name.

The new logic detects several scenarios like:
- a backup exists, but no marker exists (meaning it completed)
- backup does not exist, but marker is left behind (it gets truncated)

The marker file will be removed only when the backup completes.

This new logic does not guarantee relationships may not get duplicated.
When terminated gracefully the system should write the last page received
and close and flush the files. But if the process was abruptly terminated
(e.g. SIGKILL) it could lead to relationships being written to the OCF
file, but the marker not being updated.