print usage only on validation errors
the usage was being printed on any error, which
was rather unpleasant to look at because it makes
the output difficult to parse for clues on what happened.
The whole goal of this commit is to improve UX.

Cobra supports "SilenceErrors" and "SilenceUsage", and a callback
function "SetFlagErrorFunc" for flag parsing errors. Unfortunately,
only a subset of errors fall under "flag parsing error" category.

To amend this, a special error Wrapper is introduced that signals
an error is a validation error, so the command can manually print
the usage when "SilenceUsage" is enabled, which this commit does.

A few tests are added that assert the output. Due to the
pervasive use of globals, some refactoring was conducted in
cmd.go to make it more test friendly.