fixes potential SIGSEV when failing on the first request
since at that point we would have not received any response
and we wouldn't have a token.

Also made sure to verify the tests failed if I removed
the usage of the last known token, which led me to
improve the assertions, as the tests were SIGSEV'ing too.

Also inverted the !errors.Is(err, io.EOF) check, which
is more idiomatic